Expose endpoints for MCP
